zcash_pool_migration: rename `state` to `lifecycle`
`state.rs` did not hold `MigrationState`; it held an `impl` block on a type
declared elsewhere. The name invited exactly the wrong guess, and the split
it described was real: the data and the methods that merely read it on one
side, the queries and transitions that decide what happens next on the
other.

Extracting `model` in an earlier commit finished that split by moving the
data-only methods (`from_parts`, `funding_notes`, `crossing_values`,
`transfer_crossing_value`, `set_transaction_proved`) down with the type,
leaving this module holding only behaviour: what may be proved or broadcast
now, what has expired, what each transaction is waiting on, and the
transitions that record a broadcast, a mining, or a signature. `lifecycle`
is what that is.

The old name stays as a module alias, so `zcash_pool_migration::state::*`
keeps resolving for downstream code.

Two tests exercising `transfer_crossing_value` stay here rather than
following the method into `model`: they lean on this module's
`state_with_crossings` scaffolding, and duplicating it to move two
assertions would cost more than it clarifies.
